requirements.txt
# Core ML Libraries
# Note: Install PyTorch separately first with:
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
transformers>=4.30.0
accelerate>=0.20.0
datasets>=2.12.0
bitsandbytes>=0.39.0

# API Server
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
rich>=13.0.0
tqdm>=4.65.0

# Logging & Monitoring
loguru>=0.7.0

# Optional: Vector DB for RAG (commented out for now)
# chromadb>=0.4.0
# sentence-transformers>=2.2.0
